fix error when logout
